+++ viadot/flows/__init__.py
@@ -11,6 +11,8 @@ from .sharepoint_to_adls import SharepointToADLS
 from .cloud_for_customers_report_to_adls import CloudForCustomersReportToADLS
 from .aselite_to_adls import ASELiteToADLS
 from .bigquery_to_adls import BigQueryToADLS
+from .outlook_to_adls import OutlookToADLS
+from .salesforce_to_adls import SalesforceToADLS
 
 try:
     from .sap_to_duckdb import SAPToDuckDB
@@ -27,3 +29,4 @@ except ImportError:
     pass
 
 from .sql_server_to_duckdb import SQLServerToDuckDB
+from .epicor_to_duckdb import EpicorOrdersToDuckDB